Use new way of deployment for OSIAM
OSIAM now loads assets and configuration from a directory in the
file system known as `OSIAM_HOME`.
Add a Tomcat deployment descriptor to set the home directory for the
deployment.
Put the home directory under `/var/lib/osiam`, because it fits quite
nicely there.
This is also the home directory used in the official docs.

The files for the home directory are now under `src/test/docker`,
because they are not part of the test classpath.
They're only needed to build the Docker image.
This way the respective configuration files for postgres and mysql can
be copied by the Docker build process.
This was realized with Maven before, which led to quite a pile of
XML.
It's cleaner and easier to understand this way.

The file `greenmail-webapp-1.4.1.war` seems not to be used anywhere, so
it has been removed.